requirements.txt
tensorflow
opencv-python-headless
numpy
matplotlib
scikit-learn
streamlit
pillow
gdown
Flask
gunicorn